search/pyproject.toml
[project]
name = "search-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "bcrypt>=5.0.0",
  "black>=25.11.0",
  "dotenv>=0.9.9",
  "elasticsearch>=9.2.0",
  "fastapi[standard-no-fastapi-cloud-cli]>=0.121.2",
  "grpcio==1.71.0",
  "opentelemetry-api>=1.38.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.38.0",
  "opentelemetry-instrumentation-fastapi>=0.59b0",
  "opentelemetry-instrumentation-grpc>=0.59b0",
  "opentelemetry-instrumentation-requests>=0.59b0",
  "opentelemetry-sdk>=1.38.0",
  "pika>=1.3.2",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "protobuf==6.30.2",
  "psycopg2==2.9.10",
  "pyjwt>=2.10.1",
  "pytest>=9.0.1",
  "python-decouple>=3.8",
  "python-dotenv==1.1.0",
  "requests>=2.32.5",
  "ruff>=0.14.5",
  "sqlalchemy==2.0.40",
  "uvicorn[standard]>=0.38.0",
]
